requirements.txt
npm install expo
npm i spotify-web-api-node
pip3 install flask
Flask==3.1.0
pyopenssl==20.0.1
flask-talisman==0.7.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
flask-cors==3.0.10